requirements.txt
streamlit==1.33.0
groq==0.3.0
python-dotenv==1.0.0
langchain==0.1.16
langchain-groq==0.1.2
langchain-text-splitters==0.0.1
pypdf==4.2.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
tiktoken==0.6.0
unstructured==0.13.0